src/api/lambdas/query_index/requirements.txt
openai>=1.40.0
psycopg2-binary==2.9.9
aws_lambda_powertools
llama-index-core==0.12.0
llama-index-vector-stores-postgres==0.4.0
llama-index-embeddings-openai==0.3.0
llama-index-llms-openai==0.3.0
numpy==1.26.4
tiktoken==0.6.0
pgvector